Split testing flow into separate files
