+++ src/testing/AddTestFlow.tsx
@@ -1,30 +1,13 @@
 import React, { useState, useEffect } from 'react';
-import {
-  Message,
-  Button,
-  Text,
-  Heading,
-  Select,
-  Spinner,
-  Label,
-  Box,
-  Input,
-  Checkbox,
-} from 'theme-ui';
-import { useFormik } from 'formik';
-import { format } from 'date-fns';
+import { Heading, Select, Label } from 'theme-ui';
 
-import {
-  CreateTestCommand,
-  TestType,
-  FieldSchema,
-  ObjectSchema,
-  FilledSchema,
-  createTest,
-} from '../api';
-import { useTestTypes, useUser } from '../resources';
-import { useAuthentication } from '../authentication';
+import { CreateTestCommand, createTest } from '../api';
 import { useHistory } from 'react-router-dom';
+
+import { useTestTypes } from '../resources';
+import { useAuthentication } from '../authentication';
+import { ConfirmIdentity } from './ConfirmIdentity';
+import { AddTestForm } from './AddTestForm';
 
 export const AddTestFlow = ({ userId }: { userId: string }) => {
   const [selectedTestTypeId, setSelectedTestTypeId] = useState<string>();
@@ -77,116 +60,3 @@ export const AddTestFlow = ({ userId }: { userId: string }) => {
   );
 };
 
-const AnyBox = Box as any;
-
-const AddTestForm = ({
-  testType,
-  onComplete,
-}: {
-  testType: TestType;
-  onComplete: (command: CreateTestCommand) => any;
-}) => {
-  const form = useFormik({
-    initialValues: getInitialValues(testType.resultsSchema),
-    onSubmit(details: FilledSchema) {
-      onComplete({ testTypeId: testType.id, results: { details } });
-    },
-  });
-
-  return (
-    <>
-      <Text>Result</Text>
-      <AnyBox as="form" sx={{ display: 'grid', gridGap: 4 }} onSubmit={form.handleSubmit}>
-        {Object.entries(testType.resultsSchema.properties).map(([key, value]) => {
-          if (value.type === 'boolean') {
-            return (
-              <Box key={key}>
-                <Label>
-                  <Checkbox {...form.getFieldProps(key)} />
-                  {value.title}
-                </Label>
-              </Box>
-            );
-          }
-
-          return (
-            <Box key={key}>
-              <Label htmlFor={`test-${key}`}>{value.title}</Label>
-              <Input
-                type={value.type === 'number' ? 'number' : 'text'}
-                id={`test-${key}`}
-                {...form.getFieldProps(key)}
-              />
-            </Box>
-          );
-        })}
-        <Button type="submit" variant="block">
-          Save
-        </Button>
-      </AnyBox>
-    </>
-  );
-};
-
-function getInitialValues(schema: ObjectSchema) {
-  const initialValueEntries = Object.entries(schema.properties).map(([key, value]) => [
-    key,
-    initialPropertyValue(value),
-  ]);
-  return Object.fromEntries(initialValueEntries);
-}
-
-function initialPropertyValue({ type }: FieldSchema) {
-  switch (type) {
-    case 'boolean':
-      return false;
-    case 'number':
-      return 0;
-    case 'null':
-      return null;
-    case 'string': // fallthrough
-    default:
-      return '';
-  }
-}
-
-const ConfirmIdentity = ({
-  userId,
-  onConfirm,
-  loading,
-}: {
-  userId: string;
-  onConfirm: () => any;
-  loading: boolean;
-}) => {
-  const { user } = useUser(userId);
-  if (!user) {
-    return <Spinner mx="auto" mt={4} sx={{ display: 'block' }} />;
-  }
-
-  const { firstName, lastName, dateOfBirth } = user.profile!;
-  const { address1, address2, city, region, postcode } = user.address!;
-  const text = (value?: string) => (value ? <Text>{value}</Text> : null);
-  // TODO: icon for warning message
-  // TODO: show country here
-  return (
-    <>
-      <Message mb={3} variant="warning">
-        <Text>Confirm identity</Text>
-        <Text as="small">Check the patient's photo ID to confirm these details are correct</Text>
-      </Message>
-      <Heading as="h1" mb={3}>
-        {firstName} {lastName}
-      </Heading>
-      <Text mb={4}>Date of birth: {format(new Date(dateOfBirth), 'dd/MM/yyyy')}</Text>
-      {text(address1)}
-      {text(address2)}
-      {text(city)}
-      {text(region)}
-      {text(postcode)}
-      <Button variant="block" mt={4} onClick={onConfirm} disabled={loading}>
-        Confirm patient's identity
-      </Button>
-    </>
-  );
-};
